Reorganize code to work with standard go commands, use gofmt formatting.
